feat(challenge2): :zap: Add submission for Challenge 2
Used a pre-hook to create the stage and PUT the parquet data. Created one dbt model to store the data as a table. Created a view model to select the relevant columns, and post-hooks in this model to create the stream and perform the updates. Created a final view model to query the results of the stream